# Perf/syllable tokenize